requirements.txt
customtkinter
dxcam_cpp
mss
keyboard
pydirectinput
winrt-runtime
winrt-Windows.Graphics.Imaging
winrt-Windows.Media.Ocr
winrt-Windows.Storage.Streams
winrt-Windows.Foundation
Pillow
tomlkit
opencv-python
numpy
rapidfuzz
